requirements.txt
openai
python-dotenv
requests
streamlit
twilio
